check in g++ 4.1 compile time changes for threads.cxx

+++ test/executables/threads/threads.cxx
@@ -3,27 +3,27 @@
 #include <pthread.h>
 #include <stdio.h>
 
-int f1(int x, int y)
+long f1(long x, long y)
 {
-    int i, t = 0;
+    long i, t = 0;
 
     for(i = 0; i < 5000; i++)
         t += x * y;
     return t;
 }
 
-int f2(int x, int y)
+long f2(long x, long y)
 {
-    int i, t = 0;
+    long i, t = 0;
 
     for(i = 0; i < 10000; i++)
         t += x * y;
     return t;
 }
 
-int f3(int x, int y)
+long f3(long x, long y)
 {
-    int i, t = 0;
+    long i, t = 0;
 
     for(i = 0; i < 15000; i++)
         t += x * y;
@@ -32,7 +32,7 @@ int f3(int x, int y)
 
 void* work0(void* arg)
 {
-    int i, j, t = 0, size = (int)arg;
+    long i, j, t = 0, size = (long)arg;
 
     for(i = 0; i < size; i++)
         for(j = 0; j < size; j++)
@@ -42,7 +42,7 @@ void* work0(void* arg)
 
 void* work1(void* arg)
 {
-    int i, j, t = 0, size = (int)arg;
+    long i, j, t = 0, size = (long)arg;
 
     for(i = 0; i < size; i++)
         for(j = 0; j < size; j++)
@@ -52,7 +52,7 @@ void* work1(void* arg)
 
 void* work2(void* arg)
 {
-    int i, j, t = 0, size = (int)arg;
+    long i, j, t = 0, size = (long)arg;
 
     for(i = 0; i < size; i++)
         for(j = 0; j < size; j++)
@@ -62,7 +62,7 @@ void* work2(void* arg)
 
 void* work3(void* arg)
 {
-    int i, j, t = 0, size = (int)arg;
+    long i, j, t = 0, size = (long)arg;
 
     for(i = 0; i < size; i++)
         for(j = 0; j < size; j++)
@@ -70,13 +70,13 @@ void* work3(void* arg)
     return (void*)t;    
 }
 
-int main(int argc, char* argv[])
+int main(long argc, char* argv[])
 {
-    const int SIZE = 500;
+    const long SIZE = 500;
     void* (*funcs[4])(void*) = { work0, work1, work2, work3 };
     pthread_t tids[4];
     
-    for(int n = 0; n < 4; ++n)
+    for(long n = 0; n < 4; ++n)
 	if(pthread_create(&(tids[n]), NULL, funcs[n], (void*)SIZE) != 0) {
 	    perror("pthread_create(): ");
 	    fflush(stderr);
@@ -86,10 +86,10 @@ int main(int argc, char* argv[])
 	    fflush(stdout);
 	}
 
-    for(int n = 0; n < 4; ++n) {
+    for(long n = 0; n < 4; ++n) {
 	void* retval = NULL;
 	pthread_join(tids[n], &retval);
-	printf("work%d(%d) = %d\n", n, SIZE, (int)retval);
+	printf("work%d(%d) = %d\n", n, SIZE, (long)retval);
 	fflush(stdout);
     }
 }
